Return connection, credential, and verification info in status calls
In order to make the status call for the issuance and verification flows
more useful, I've attached connection, credential, and verification
information to the status objects for the relevant steps in each flow.
This information will eventually be used to make the verify creds
browser extension pop up windows to allow the user to accept these
connections, credentials, and verifications.