Mitigate XSS by blocking inline scripts
First, I had to replace the inline script that stores the logged in user's
id in the window with a cookie.  Then, I updated the `Content-Security-Policy`
header to block all inline scripts and and scripts loaded from unknown
sources.  I'm sure this isn't really bullet proof, but this blocks the test
cases that I've been given, and I believe it meets a basic level of security
that should protect developers using these samples to build their own PoC
issuer pages.